backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
sentence-transformers
chromadb
torch
transformers
pytesseract
pillow
pypdf
python-multipart
numpy
pydantic
requests
